Move sortedness utilities to `order-sortedness.c`
